+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 ### Changed
+- Add support for overloaded contract functions (PR #166)
 - Removed `error_chain` and reworked errors manually. Also removes the `backtrace` feature. (#167)
 
 ## [10.0.0] - 2020-01-08